+++ voxygen/i18n-helpers/src/lib.rs
@@ -109,7 +109,7 @@ pub fn localize_chat_message(
                 BuffKind::Regeneration
                 | BuffKind::Saturation
                 | BuffKind::Potion
-                | BuffKind::Swiftness
+                | BuffKind::Retreating
                 | BuffKind::CampfireHeal
                 | BuffKind::EnergyRegen
                 | BuffKind::IncreaseMaxEnergy