+++ CHANGELOG.md
@@ -9,7 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Swiftness potion
 - Esperanto translation
 - Item quantity sort in player inventory.
 - Using Block('Alt' by default) in Defensive Stance now feels stronger
@@ -39,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A way to target non-player entities with commands. With rtsim_id: `rtsim@<id>`, with uid: `uid@<id>`.
 - Shorthand in voxygen for specific entities in commands, some examples `@target`, `@mount`, `@viewpoint`.
 - Added hit_timing to BasicMelee abilities
+- Retreating Potion
 
 ### Changed